Allow setting a different destination file name from local file name in chunked upload
Using macOS File Provider APIs we rely on the system to provide us with
a URL pointing the an item's local modified contents. This URL last path
component is a UUID and does not represent the expected file name of the
item.

The current API of uploadChunk presents a problem because it assumes the
local content file's filename is the same as what will eventually be
uploaded to the server. This commit addresses the issue by allowing
users of this function to provide a destination file name which will
replace the provided local filename, if used.

Signed-off-by: Claudio Cambra <claudio.cambra@nextcloud.com>