Allow setting custom destinations in chunked upload (#116)
* Allow setting a different destination file name from local file name in chunked upload

Using macOS File Provider APIs we rely on the system to provide us with
a URL pointing the an item's local modified contents. This URL last path
component is a UUID and does not represent the expected file name of the
item.

The current API of uploadChunk presents a problem because it assumes the
local content file's filename is the same as what will eventually be
uploaded to the server. This commit addresses the issue by allowing
users of this function to provide a destination file name which will
replace the provided local filename, if used.

Signed-off-by: Claudio Cambra <claudio.cambra@nextcloud.com>

* Make file chunks output directory customisable

Currently the chunking procedure produces file chunks within the
directory of the input file. However, this fails in cases where the
directory is read-only.

This can be fixed by allowing the chunked files output directory to be
changed

Signed-off-by: Claudio Cambra <claudio.cambra@nextcloud.com>

---------

Signed-off-by: Claudio Cambra <claudio.cambra@nextcloud.com>